fixed some lint warnings
